# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DOCSWAP is a platform designed for realtors to efficiently share, manage, and analyze documents on a neighborhood level. It enables realtors to make better use of the content created for each listing by organizing it into a searchable database that supports both uploads and downloads of documents. These documents are categorized to facilitate market analysis, putting the power of data directly into the hands of realtors and market analysts interested in housing information in the lower mainland.",
        "title": "DOCSWAP API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "DOC-SWAP Support",
            "url": "http://www.doc-swap.com/support",
            "email": "support@doc-swap.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/category": {
            "get": {
                "description": "get all categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Retrieve all categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Category"
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/category/search": {
            "post": {
                "description": "Search for categories with the given search parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Search categories",
                "parameters": [
                    {
                        "description": "Search object",
                        "name": "searchObj",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/search.Search"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the category",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Category"
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/category/{id}": {
            "get": {
                "description": "get a category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Retrieve a category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Category"
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Category": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent": {
                    "$ref": "#/definitions/models.Category"
                },
                "parentID": {
                    "type": "integer"
                }
            }
        },
...
        "search.LogicalOperator": {
            "type": "string",
            "enum": [
                "AND",
                "OR"
            ],
            "x-enum-varnames": [
                "And",
                "Or"
            ]
        },
        "search.Operator": {
            "type": "string",
            "enum": [
                "=",
                "!=",
                "\u003e",
                "\u003e=",
                "\u003c",
                "\u003c=",
                "LIKE",
                "NOT LIKE",
                "IN",
                "NOT IN",
                "IS NULL",
                "IS NOT NULL"
            ],
            "x-enum-varnames": [
                "Equal",
                "NotEqual",
                "GreaterThan",
                "GreaterThanOrEqual",
                "LessThan",
                "LessThanOrEqual",
                "Like",
                "NotLike",
                "In",
                "NotIn",
                "IsNull",
                "IsNotNull"
            ]
        },
        "search.Param": {
            "type": "object",
            "properties": {
                "AssociationForeignKey": {
                    "type": "string"
                },
                "Field": {
                    "type": "string"
                },
                "Operator": {
                    "$ref": "#/definitions/search.Operator"
                },
                "Value": {}
            }
        },
        "search.Search": {
            "type": "object",
            "properties": {
                "LogicalOperator": {
                    "$ref": "#/definitions/search.LogicalOperator"
                },
                "Params": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/search.Param"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}